Rework Line animation functions to be more generic.
