+++ packages/ag-charts-community/src/chart/series/cartesian/areaUtil.ts
@@ -4,15 +4,9 @@ import type { Path } from '../../../scene/shape/path';
 import type { ProcessedOutputDiff } from '../../data/dataModel';
 import type { SeriesNodeDatum } from '../seriesTypes';
 import type { CartesianSeriesNodeDataContext, CartesianSeriesNodeDatum } from './cartesianSeries';
-import {
-    pairCategoryData,
-    pairContinuousData,
-    prepareLineMarkerAnimation,
-    prepareLinePathAnimationFns,
-    renderPartialLine,
-} from './lineUtil';
-import type { MarkerChange } from './markerUtil';
-import type { PathPoint } from './pathUtil';
+import { pairCategoryData, pairContinuousData, prepareLinePathAnimationFns, renderPartialLine } from './lineUtil';
+import { prepareMarkerAnimation } from './markerUtil';
+import type { PathPoint, PathPointChange } from './pathUtil';
 
 export enum AreaSeriesTag {
     Fill,
@@ -67,11 +61,11 @@ export interface AreaSeriesNodeDataContext
     strokeData: AreaPathDatum;
 }
 
-function renderPartialArea(pairData: PathPoint[], ratios: Partial<Record<MarkerChange, number>>, path: Path) {
+function renderPartialArea(pairData: PathPoint[], ratios: Partial<Record<PathPointChange, number>>, path: Path) {
     const { path: areaPath } = path;
     let previousTo: PathPoint['to'];
     for (const data of pairData) {
-        const ratio = ratios[data.marker];
+        const ratio = ratios[data.change];
         if (ratio == null) continue;
 
         const { from, to } = data;
@@ -135,6 +129,24 @@ function pairFillContinuousData(newData: AreaSeriesNodeDataContext, oldData: Are
     };
 }
 
+function pairStrokeCategoryData(
+    newData: AreaSeriesNodeDataContext,
+    oldData: AreaSeriesNodeDataContext,
+    diff: ProcessedOutputDiff
+) {
+    const oldPoints = { scales: oldData.scales, nodeData: oldData.strokeData.points };
+    const newPoints = { scales: newData.scales, nodeData: newData.strokeData.points };
+
+    return pairCategoryData(newPoints, oldPoints, diff);
+}
+
+function pairStrokeContinuousData(newData: AreaSeriesNodeDataContext, oldData: AreaSeriesNodeDataContext) {
+    const oldPoints = { scales: oldData.scales, nodeData: oldData.strokeData.points };
+    const newPoints = { scales: newData.scales, nodeData: newData.strokeData.points };
+
+    return pairContinuousData(newPoints, oldPoints);
+}
+
 export function prepareAreaPathAnimation(
     newData: AreaSeriesNodeDataContext,
     oldData: AreaSeriesNodeDataContext,
@@ -149,6 +161,13 @@ export function prepareAreaPathAnimation(
         return pairContinuousData(newData, oldData);
     };
 
+    const prepareStrokePairs = () => {
+        if (isCategoryBased && diff) {
+            return pairStrokeCategoryData(newData, oldData, diff);
+        }
+        return pairStrokeContinuousData(newData, oldData);
+    };
+
     const prepareFillPairs = () => {
         if (isCategoryBased && diff) {
             return pairFillCategoryData(newData, oldData, diff);
@@ -157,15 +176,21 @@ export function prepareAreaPathAnimation(
     };
 
     const { resultMap: markerPairMap } = prepareMarkerPairs();
+    const { result: strokeResult } = prepareStrokePairs();
     const { top, bottom } = prepareFillPairs();
 
-    if (markerPairMap === undefined || top.result === undefined || bottom.result === undefined) {
+    if (
+        strokeResult === undefined ||
+        markerPairMap === undefined ||
+        top.result === undefined ||
+        bottom.result === undefined
+    ) {
         return;
     }
 
     const pairData = [...top.result, ...bottom.result.reverse()];
     const fill = prepareLinePathAnimationFns(newData, oldData, pairData, renderPartialArea);
-    const stroke = prepareLinePathAnimationFns(newData, oldData, top.result, renderPartialLine);
-    const marker = prepareLineMarkerAnimation(markerPairMap);
+    const stroke = prepareLinePathAnimationFns(newData, oldData, strokeResult, renderPartialLine);
+    const marker = prepareMarkerAnimation(markerPairMap);
     return { fill, marker, stroke };
 }
